Rather nicely, pypy supports arbitrary depth call stacks.
Unfortunately the support for recursive structures currently relies on
an exception being thrown when the recursion level gets too deep. This
not happening means that pypy will cheerfully eat all the memory on the
system in a matter of seconds. This is sad.

For now, disable this test under pypy until we've got a better solution.