Add a strategy that produces Random values.
The major purpose of this is so that you can test randomized algorithms
in a reproducible manner - this allows Hypothesis's minimization to work
sensible in those cases.

In order to get slightly better behaviour with printing and such-like
we actually return a subclass of Random which displays its original seed.
This makes the problems much easier to understand and reproduce.

This commit also creates a recursive self test for copying to include a
provided Random as an additional test of this behaviour.