Statistical correction for multiple testing.
This changes the individual tests for statistical distributions to
have a much stricter p value for rejection but adds a final test at
the end which adjusts for multiple testing to see if any of them are
significant. It also adjusts the probability requirements to ones that
better match the current actual distribution.

In practice this is all a bit of gilding the lily. The nice thing about
running statistical tests against something where you control the data
generation is you can trivially run enough that basically any test
conclusion you get is going to be statistically significant at whatever
level you want it to be at. The source of the error is never going to
be your p value, it's going to be the assumptions you didn't realise
you were making when you wrote the test.

But the Benjamini-Hochberg procedure is cool, so I'm going to leave
this in anyway.