Change the distribution of bounded float ranges
The reason for this is because one of the tests is proving flaky
and this is evidence that hypothesis isn't covering those regions
very well. This distribution allows it to get better coverage of
certain shapes, as reflected by the increased probabilities in
test_statistical_distributions.

Hopefully this will both improve coverage in general and also make
the build less flaky.