Fix a bug in defining strategies on subtypes
If you had a pattern of usage that looks like this:

    mapper = SpecificationMapper()
    mapper.define_specification_for_instances(BBA, const(0))
    mapper.define_specification_for_instances(A, const(1))
    mapper.define_specification_for_instances(BBB, const(2))
    assert mapper.specification_for(BBB()) == 2

This would *sometimes* produce the wrong answer and sometimes not,
depending on the vagaries of dict iteration order (which is
randomized for each process I think but haven't checked). Essentially
the culprit was an entirely bogus implementation of topologically
sorting by the subclass relationship. There is now a non-bogus
implementation.