Remove dependency on six
I realised that hypothesis used really very little of six and that six
itself was quite a complex beast. So it took very little work to create
a small shim that had all the bits we need for python 3 compatibility.

This in turn makes it easier for people to depend on hypothesis because
there's no concern about six versions. This probably never was a concern
because honestly six seems pretty good about version compatibility, but
it's nice to remove the dependency.