Don't count recursive tests for coverage
This is for a mix of reasons, but mostly that I'm a little suspicious
of the quality of testing they provide - they're good for finding
interesting obscure edge cases but I'd rather not count on them for
core testing.

They also run with very restricted size and number of examples for
performance reasons (they basically run in O(normal performance^2)
because they're running hypothesis inside hypothesis) so their ability
to discover tests is much less solid than running tests directly.

This has the novel feature of being one of the first significant
changes I made to Hypothesis's test suite which *didn't* uncover a
new bug.