Reuse previously saved examples
This is currently a bit of a bodge and in all honest this entire
method needs a rethink because it's nasty and has grown far too
complicated.

In particular the way this works at the moment does *not* play
well with the timeout and max_examples features in the sense that
it completely ignores them. In the current usage pattern this is
actually totally fine because the database is always going to be
small, but it will not be fine once e.g. interestingness detection
lands.

Anyway, consider this a proof of concept rather than the version of
this code that is going to land in a production version, but it
does work!