Provide an opt out method for serialization
Serializing descriptors would require either

  a) A full without-specific serialization protocol which I didn't want to do
  b) Using pickle, which is horrendously insecure and nope.

So I didn't really want to do it. On the assumption that other people
will run into this problem as well I made it a general mechanism. A
type can explicitly mark itself as not serializable, at which point the
verifier won't try to serialize it.

This does of course mean you won't be able to reuse previous examples. Use
this feature with caution.