Print falsifying examples on failure
I've been consistently finding that py.test doesn't actually do a
very good job of printing out the falsifying example and that I kept
needing to add manual prints in to figure out what was going on. If
this is happening to me it's probably happening to other users too.

So, we now print the falsifying example if we find one. We try to
use keyword arguments where possible but will fall back to positional
ones in the presence of varargs (not that varargs are properly supported
in given yet, but for when they are...)

Note: Printing inside a library is a bit annoying but I think in this
case justified. When a general reporting mechanism is available this
should be subsumed into that.