Begin the process of renaming.
Basically the plan is to change the current Format to Converter and
to repurpose the name Format to refer to "json" in this case. The
goal is to make it easier to store their examples in e.g. msgpack if
they want to and have the converter responsible for just marshalling
to and from basic object representations.