Converter for StatefulTests.
Serialize steps by name rather than index because this is something
that is likely to change a lot. Otherwise mostly do the obvious thing
with serializing arguments according to their defined conversions.

Note that this is currently very poorly covered by tests and will
fail for that reaosn, but I've verified that test_statefultesting
runs correctly with the DB on.