Better selection of new parameters
Idea: We track whether the last parameter was new, and thus we track
the number of times a new parameter has produced a good result. This
lets the "generate a new parameter and pick from it" arm participate
in the bandit like any other, so that we pick it when it seems like
the best bet based on prior evidence.