.pyc files are the bane of my existence
OK, so, I'm not 100% sure this is the problem, but this fixes the issue
locally.

At *some* point in the loop running concurrent versions of the tests in
different versions of python gets things really fucked up. It's either an
intrinsic problem with how Travis is running these or the result of pytest's
approach using bytecode rewriting.

Either way pyc files are the devil and must be purged from our community with
blood, fire and environment variables.