Raise expected number of samples from small sets
The reason the BadSetTester test was so flaky was because it was
spending a lot of time just generating examples which only used one
method. Don't do that then.

This feels like over fitting but I think it's a valid decision. It
is generally not a great idea to generate single elements sets, but
we do want to do so occasionally. Still, a more principled choice
would be nice.