Save subexamples in the database
The idea is that when we have discovered an interesting example
then bits of it are probably interesting too. This isn't necessarily
true but is generally worth checking out. So we save those values
in the database too.

This will be much more interesting once we can also get those values
back out of the database as bits of other things.