Augment strategies with values from the database
Idea: If there are values in the database for a descriptor then
sometimes we should use those.

This allows us to reuse parts of other examples we've found in other
tests in building up our new examples. e.g. if the number 42 has turned
out to be interesting then as awell as reusing that for int examples
we can also construct the list [42] more readily, or the tuple (42, 3)
(or even the tuple (42, 42))

This also makes preseeding the database with interesting values a much
more attractive proposition. Previously because all the descriptors
we store in the database are of the form ((args), {kwargs}) it would
have been rather awkward.