Significantly expanded support for floating point
Generate floats in a much larger variety than previously encountered

In particular now will generate:

  1. NaN
  2. Infinite values
  3. Finite integer values which will be too large to represent exactly
  4. Really small non-zero values

This also fixes a number of issues that would previously have given you
a really hard time if you'd tried to do this yourself: A lot of Hypothesis's
internals did not behave at all well if you passed it a value that is not
equal to itself.

This will actually still cause a bunch of problems in the general case
unfortunately, but the special case of NaN is handled. For the general case
until someone comes up with a really good use case I'm going to adopt a
policy of "Well don't do that then".