Better support for custom repr and Just
Previously we'd always use repr for Just. This is wrong behaviour.
In particular this was breaking now because we don't want to always
use repr for float and complex.

This is wrong, but we still want to support custom reprs.

Solution: We will *always* use your custom repr unless you're one
of the types we have special cased or you are inheriting the repr
from object.