SpecMapper inheritance of defaults was bad
If you called SpecificationMapper.default() then this would become
the default for all subclasses if they hadn't already been called.

I'm not sure what the correct semantics for this *should* be, but
it's definitely not that.

I don't think this ever caused any problems because of how internal
this thing is, but ugh.