Don't collapse not actually_equal descriptors
This was causing a *really* annoying Heisenbug in some of the tests.

The problem is that if you have something like a strategy for one_of((
{int}, frozenset({int}))) this would have reported its descriptor as
{int} because the two elements are equal. This is Bad because a strategy
you get from {int} cannot produce a frozenset({int}) and this was causing
problems when round-tripping strategy -> descriptor -> strategy deep in
the bowels of some of the tests.