Simultaneous simplification of tuples
The test test_searchstrategy.test_can_simplify_named_tuples would
sometimes take an arbitrarily long time to simplify. This turned
out to be because the way it worked it was actually quite hard for
it to converge on a good simplification because it would try
simplifying just the first coordinate, then the second, etc.

This changes the simplification order to try to mix up the different
options. Instead of just simplifying the first coordinate, then the
second, etc. it round-robins between the options. This can produce
much faster convergence in some circumstances such as this.