Provide a standard test suite for implementors
This allows people who want to implement their own strategies to
run a standard test suite that validates that they've done it right
with only a single line of code in their normal tests.

Obviously more specific tests may be required, but this does a pretty
good job of catching a lot of the obvious gotchas.