Handle Just values which are not deep copyable
This is needed because pytest Session objects are not deep copyable,
but is probably a good idea in general.

It does of course mean that when passed mutable things that are not
deepcopyable we might get flaky errors but hopefully those should be
caught by the flaky error detection. If not, well don't do that then.