Initial model introspection and generation
This introduces the extreme basics of automatically generating your
Django models for you.

It currently does not support the full range of fields and will not
work if you have non-nullable foreign keys. It also does no simplification.
It should be considered a sketch of the end result rather than something
that is ready for use.

But it works! Yay!