Major API redesign and rewrite
This commit is basically about taking the separation between templates
and values seriously. Templates now look much less like the values they reify
to. This allows us to drop a bunch of cruft. In particular could_have_produced
is now dead: It's a logical error to pass a template to a strategy that did not
produce it, and we can encode additional information into the template to make
sure that this invariant is preserved.

In the course of doing so I realised that because templates were now such
an internal facet of strategies that it didn't make sense to have converters
separate from strategies, and keeping them separate was creating lots of hairy
bugs that were hard to track down, so I merged the two.

This work isn't even close to done but this commit was large enough as it is.
This basically gets it to the point where the main tests pass without running
in database mode. I expect I've lost a lot of coverage (many tests were deleted
because they don't make any sense now), database mode is unlikely to be fully
operational, and I haven't even started fixing the hypothesis-extra packages.