Split up strategytable into a bunch of submodules.
The file was getting ridiculously huge. This breaks it up into
a bunch of smaller files.

This isn't really intended as a semantic change, though it does
have the nice side effect of giving searchstrategy a much better
defined API

This commit also moves where the definitions of mapping strategies
to descriptors happens to the individual modules where the strategies
are designed. This is partly for consistency with how things work in
hypothesis-extra, partly as prep for future refactoring of how the
whole StrategyTable stuff works.