Significantly more elaborate settings
The goal here is to allow strategies to define settings and to use them
in setting themselves up. This will let verifiers work with smaller
strategies if desired.

Right now though this is just a refactoring which achieves more or less
the same behaviour as before but with more code.

Conflicts:

	src/hypothesis/internal/specmapper.py