Wrap the Random we use in a BuildContext type.
The idea is that the BuildContext is going to be able to store previously
built templates. We'll then generate values in a way that stores these on
the build context and reuses them. This can then be prepopulated with stuff
from the database.

Nicely it also gets us a bunch of what was nice about testmachine: Value
reuse within examples.