That which can be killed by refactoring should be.
Remove the ubiquitous .descriptor from SearchStrategy. It turned
out to only be used for repr (and was often misleading when it was!)
and supporting it a) required a bunch of unneccessary bookkeeping and
b) Made the interface to map more annoying.

This means that Hypothesis strategies can now actually be monadic! Yay.

(I haven't actually implemented bind, but it would now be doable.
Previously the weird second argument to map would have been required
there too)