Do joint simplification of identical elements
Idea: If we have a list containing the same element twice some tests
will only fail if those values are identical. We'd still like to be
able to replace them with a simpler identical value.

I've been apologising for the lack of this feature for ages, so it's
a little embarrassing how easy it was to add. The reason for that is
that the template approach makes this much easier because we can
guarantee that types are hashable and that object identity is not
a problem (e.g. previously if we'd been simplifying two lists that
were value but not reference equal we'd have to worry about copying
them so that they stayed that way).