Don't use an easily shadowed name
This fixes issue #38. The problem was that the argument to accept
gets shadowed by an argument to our function so we try to call the
wrong value.

In theory you can still trigger this bug by calling an argument to
your function hypothesis_reflection_internal_name_do_not_use but
you basically have it coming if you do that and I think I'm happy
to say that part of the spec to using @given is don't do that then.