Start to modularise docs
I was given the very valid feedback that the docs as they currently
are are very poorly structured and as a result feel a bit incoherent.

The goal of this work is to split the large document up into self
contained pieces each of which cover a certain use-case for Hypothesis.

The current WIP here is that the index, quick start, community and
internals pages are about right, but everything else is a bit of a
mess and is still in need of rationalising and editing.