Additional work on simplification of lists
Add some new examples and try a new ordering of simplify
operations which makes them simplify much much faster.

The choice of examples is a bit arbitrary, but this reordering
makes a lot of sense: It frontloads on simpler operations and
operations that are likely to converge to a good answer faster.