Correctly implement floats_in_range.
The positioning of the cut was entirely off. It was intended to
indicate how far along between the two you were as a fraction of the
length but was being interpreted as an absolute length.

I then proceeded to only test it for values in (0, 1), where it would
have worked. Exactly the sort of mistake Hypothesis is meant to help
you avoid. Doh.