Use pyenv to manage python dependencies everywhere
Previously it was only used for OSX builds. This turns it on for
Linux builds too.

The original reason for this was to be able to test all versions of
Python supported by pyenv on every platform (I previously didn't have
access to pypy3 on Linux).

However this enabled a bunch of cool things: Because all my python
installs are local now, I no longer need sudo. *this* means that I
can use the new Travis container based infrastructure, and this means
that I can turn on caching!

(Turning on caching is actually non-optional for this. On Linux pyenv
builds from source which takes a horrendous amount of time to do it on
each build. By using caching we only have to pay that cost once).

In particular by caching the ~/.cache/pip directory the build becomes
much more robust against pip failures because we can always install
from cache (I haven't tested this in anger so it may be that I need
to do something to make this work a bit better, but it paves the way
for it being possible)

As a side-benefit, no longer doing apt-get update means my build can
no longer be made to fail by rabbitmq.com being slow to respond and
causing a timeout.

This commit also removes rst-lint. It wasn't proving very useful and
violates my "don't depend on things that don't work on Python 3"
policy. As a result this commit also upgrades the version the lint
process runs on to a modern version of Python. This has little
effect except to make me feel better.