New length strategies for lists
Basically we try quite hard to throw away half of the list. This
is greatly aided by judicious use of randomness.

This strongly supports my growing thesis that the right API for
simplify functions includes a source of randomness. For now it's
not the end of the world that we just seed it deterministically but
it's suboptimal.