Ad hoc bodges, err, heuristics for simplification
The main interesting feature of this commit is that it rerereintroduces
an idea I keep playing with and taking out, which is the ability to say
which of two values are simpler.

This in turn gives us a number of ways of splitting and manipulating
the data which can sometimes help, or at least helps in these particular
fairly artificial examples and at least plausible helps in others.

This branch is no longer suitable to merge until I properly sort out the
use of randomization in simplify because it's obviously needed and this
seeding it with data thing is just a waste of everybody's time.

All that being said, there are some interesting bits in here. In particular
the example cloning thing works really well when it hits, because it produces
vastly simpler examples than you'd other get which *also* feed in really well
into the simultaneous simplification. If you're lucky you hit a fast path
which shortcuts you through whole chunks of the minimize graph you just
get to completely ignore.

The pivoting method also appears to be useful. It's based on the observation
that many errors are only triggered by a set of some sufficiently complex
examples, so by attempting to throw away all sufficiently complex or all
sufficiently simple examples you might get a win.

All of this may just be overfitting which isn't of any huge amount of
use in the general case, but on the artificial example set all of these
together have made the run times much more consistent and generally faster.

The less said about the integer simplification heuristics the better.