Delete searchstrategy tests.
They're pretty bit-rotted and have been patched a lot over the commits,
and they increasingly serve little useful purpose given both the
standard strategy tests and the quality tests. The CI will fail until
I've written new and better tests for any lines that are now uncovered
as a result of this.