Brace yourself: Monads are coming
This adds an initial sketch of making flatmap (monadic bind) work.
Basically it lets you write strategies that depend on other strategies
- you draw a value, then you build a strategy from the value, and you
draw from that.

Simplification is not supported yet, but I know how to make it work.

This involves a lot of slightly fiddly tracking of different types of
state a template can be in - it could either have come from generation,
simplification, or the database, and we need to treat these all a bit
differently.