Simplifiers now accepts a template argument
This was to support flatmap - without knowing the target strategy
you can't actually decide how to simplify the right hand value. It
turned out to be a pretty good change though which made a bunch of
things easier.

The intended behaviour is that the template argument is just a hint.
You use it to return a better set of simplifiers than you otherwise
could have, but they need to be valid for all types. Note "valid"
does not actually mean "does anything useful".

This doesn't yet support flatmap, but it's the work required to
ensure that it's possible.