Better validation of @given arguments
Hypothesis previously raised a TypeError for bad argument to given.
Unfortunately this makes it hard to test for validation because
other things can raise TypeError too. This commit introduces an
InvalidArgument subclass of TypeError which Hypothesis uses for
these circumstances. This allows the tests to actually catch the
bug in formatting that happened when a wrong keyword argument was
passed.

Closes #54.