Define an infinite stream type and strategy for it
Core idea: Infinite stream type, backed by a generator. In our case
that generator is just values drawn repeatedly from a strategy.

We can simplify this and serialize this because we have the original
seed to recreate the original stream and we can track how much has been
changed because we're terrible people who are prepared to open up the
internals of the lazy evaluation and see what's going on. Yay us.

We even get adaptive assumption by inheriting the parameter of the
underlying type (we use the usual trick of saving a parameter as the
seed that generated it).

The result is: Infinite streams with the full set of features Hypothesis
normally offers. Yay!