Examples from database must participate in timeout
This would cause a bug where slow tests with a lot of examples in
the database would run for longger than the timeout even when they
could reasonably not have.

This manifested as the timeout tests in testdecorators failing when
run as part of the database tests, but that wasn't a good test for
this problem so this commit also introduces a specific test for
this scenario.