Simplify should not increase complexity.
Previously there was no shared invariant between strictly_simpler
and simplify. They were treated as somewhat orthogonal. This wasn't
great.

This adds a fairly weak invariant: It should not be the case that
x simplifies to y and x is strictly simpler than y. This turns out
not to have been satisfied for numeric types, partly because the
strictly_simpler implementation for int was just plain wrong.