Keep multiple serialized templates around
The idea is that when we have inconsistent strategies at an index
it will *usually* be from a small set. Thus the correct thing to do
is that when we encounter bad data we don't throw it away, but we
save it for the next time something like this comes up.

In particular this means we don't throw away all that work we did
last time in simplifying our data and allows us to achieve much
better simplifications.