Introduce "warmup" rounds.
The idea is that we often find ourselves in a position where there
are some simplifiers that early on just won't ever work but we spend
a lot of time checking the. This lets us spend a few passes finding
simplifiers that will work quite well in these circumstances and
will chew the example down into something more manageable before we
start to get serious.