Add fixed point combinator for strategies
The idea is to support recursive data in a way that manages depth by
controlling the maximum number of "leaf" nodes. This lets us adapt to
any level of branching the user wants to throw at us without having to
explicitly set a branching rate.

The idea is that you specify a base state and a recursive case which can
expand any strategy to produce a new one, you specify a limit, and then
Hypothesis gives you something which will turn that into a strategy for
a tree of strategies built off that.