Better deduplication of active morphers
In general a morpher makes most sense deduplicating as its
current template, because that's what it is most likely to become
the next time you ask for something.

In practice this may not always give the right answer, but it
definitely gives the right answer in the current usage pattern
with flatmap, and tracking is inherently somewhat heuristic in the
first place.