Make: The technology of the future!
People have had a lot of annoyance developing Hypothesis if they're
not me, and part of this is a lack of a uniform interface to a lot of
the random operations in it, and another part is having to juggle the
thousands of different python versions needed.

This unifies all the common operations of Hypothesis development into
a single makefile, which is both the CI environment and the common
development environment. It manages local installs of all the relevant
pythons, so it doesn't matter if you don't have a particular version
installed.

This also includes some new organisation of tests to provide a check-fast
operation which runs a decent set of tests in only a couple of minutes
instead of the multi hour fiasco that is a full build. This required
a bit of refactoring of test suites and highlighted a few flaky tests,
so this contains a few modifications to those.

Along the way, this caused me to finally figure out what was wrong with
isort that caused it to produce different answers sometimes: It was treating
any libraries not installed as being part of "your" code, so it wasn't
sorting them in the same location. The fix for that is currently commented
out so as to not make this diff ridiculous.