Remove copying behaviour from profiles
Because the existing encouraged idiom (the only one that historically
worked properly!) for changing default values to settings is:

  Settings.default.blah = blah

The use of the mutability of settings is quite common. Unfortunately
by copying settings this produces very surprising behaviour:

  Settings.load_profile("default")
  Settings.default.blah = 1
  Settings.load_profile("otherthing")
  Settings.load_profile("default")
  assert Settings.default.blah == 1 # This is no longer true.

In particular this turned out to have surprising consequences for the
Hypothesis tests themselves!

While arguably the correct solution is "Settings should be immutable", I
think for now "don't copy the Settings object" interacts better with
existing idioms.

Note: Requires discussion with @Bachmann1234 before merge.