Implement notion of capturing a BuildContext
For things that use find() and strategies that require lazy loading
like streaming() we run into the problem that the BuildContext may
be closed but still needed if you try to do something like reading
past the part of the stream that was used.

capturing allows us to extend the lifetime of a BuildContext. When
we use find(), the last call will not close the build context until
its __del__ is called.