Initial attempts to conjecturize Hypothesis
This is half in pieces and doesn't work very well yet. Some things
are working. Most aren't. It's really an initial proof of concept
more than it is even heading in the direction of a finished form.

This commit squashes a whole bunch of work, much of which turned
out to be a bad idea, in a desperate attempt to rebase this onto
Hypothesis 2.0