Don't let health check timing disturb determinism
The problem: The health check involves a timeout. This means that
its exact use of the PRNG can depend on test timing. This in turn
can affect the determinism of the subsequent test. It shouldn't
do that.

The solution is to split the PRNG and use that
for the health check's testing. That way regardless of how many
runs the health check does or does not perform, it always leaves
the original PRNG in the same state.