Handle BuildContext and capture in find correctly
Things like choices and streaming need to be able to keep drawing
test data indefinitely once they've been found, so we provide them
with an additional source of randomness that just reads (unshrunk)
data past the end of the buffer.

This will not work very well with things that filter. It may be
worth considering deprecating these as well.