Limit failed shrinks
This is to deal with the case where we have very few successful
shrinks available to us and keep wasting our time on failed ones.

We could add another setting here, but frankly there are too many
settings already, so we just treat the allowable number of failed
shrinks as 10x the allowable number of successful ones.