Imrpove guarantes of lexicographic minimizer
In the course of other investigation I noticed that we didn't always
find the smallest integer >= i. This fixes that.

In theory most of this isn't needed to fix that, but in practice it
helps because the "try the predecessor" operation has the potential
to make everything much worse performance wise so we try extra hard
at the beginning to shrink the size of the buffer's non-zero segment
first.