Improved shrinking
Idea: We add supplementary intervals for pairs of "adjacent"
intervals. Two intervals are adjacent if:

1. One starts immediately after the other
2. They are at the same "level" in the tree structure

This keeps us from going quadratic because for every interval
there are at most two others it can be merged with.

This lets us detect cases like the list construction automatically,
but it seems to also help in other unexpected things - e.g. the
recursive examples got more reliable after this.