Support for Windows development using Vagrant
This introduces a Vagrantfile that produces a VM suitable for running
the normal Hypothesis make based build system. The intent is not that
it "deploys" Hypothesis but that vagrant ssh will give you a useful
machine to work on.

This required a few changes to the existing system:

1. The .runtimes directory is moved out of the current working
   directory and into $HOME. This is because it does not seem to
   behave at all well with living on a shared file system between
   Windows and Linux.
2. Shell scripts must never have crlf style line endings in them,
   so a .gitattributes now enforces that.

This also adds .gitignores for vagrant and idea files.