Support for Windows development using Vagrant
This introduces a Vagrantfile that produces a VM suitable for running
the normal Hypothesis make based build system. The intent is not that
it "deploys" Hypothesis but that vagrant ssh will give you a useful
machine to work on.

This required a few changes to the existing system:

1. The .runtimes and .tox directories moved out of the current working
   directory and into $HOME. This is because tools do not seem to
   behave at all well with living on a shared file system between
   Windows and Linux.
2. Require normalization of all line endings to unix style. Because the
   build is linux based it makes the most sense, and a lot of things
   break when they encounter Windows style line endings.

This also adds .gitignores for vagrant and idea files.