Improve shrinking of code using flatmap/composite
In existing Hypothesis there is a problem where often code
using these results in examples that don't shrink very well
because shrinking an earlier value results in changes to
the structure of the later value that no longer exhibits
the problematic behaviour.

This change introduces some limited ability to work past
that by allowing the later structure to be reordered in a
way that gives more flexible possibilities for producing
shrunk examples..

This is primarily focused on cases that use length prefixed
lists and similar concepts (hence the tests focusing on that),
but likely should work more generally.