Use pointer identity to skip equality check
Note that this is technically unsound: If one were to change the contents
in such a way that kept the hash value intact and ask for the same number
of items then this would produce a wrong result.

But the consequences of getting this wrong are relatively low because we
also check n_items so it will never return an out of bounds value.