Standardize which files to format. Exclude vendor
This replaces the four (!) different places we decide which files
get formatted with a single one, as a prerequisite for the actual
point of this: Formatters should not run on vendored files.

This is somewhat prompted by an intent to add another vendored
file, but it's a good change in its own right even if we don't
do that.