Fisher-Yates style implementation of permutations
This is the reverse of the normal order of Fisher-Yates because
this shrinks "correctly" (or at least more consistently with the
usual way Hypothesis does things, and the way permutations shrank
historically): Smaller is lexicographically closer to the identity
permutation.

Unlike the old implementation, this one isn't quadratic in the
size of the list. This fixes issue #438.