Deferred strategies should unwrap their arguments
We use deferred strategies all over the place. Unfortunately this
means a lot of indirection. Python is not very good at optimising
indirection.

This patch allows us to transparently ignore most of that
indirection by unwrapping all strategies before we create the
"real" underlying strategy.

Informal benchmarking suggests it helps some of the example
quality tests (which run draw() a lot of times) about 10-20%, but
this is obviously not that reliable.