Make test_terminates_shrinks more reliable
This has become slightly flaky - I've seen it fail on a few recent
builds. I think this is a mix of the new timing changes and the
new improvements to the shrinker.

This makes the example slightly harder to shrink and removes the
timeout, which should make the test more reliable.