Add min_size to test
Now that we're "running time faster" this seems to sometimes fail
at the generation stage with NoSuchExample.

Adding a min_size parameter there means that it should stop doing that.

While in the area I noticed we're now much better about shrinking this
example so I also added stronger assertions about the result.