Add a block reordering pass
This basically does a partial insertion sort on blocks of the same
length, moving cheaper blocks earlier in the process. This helps to
achieve a canonical result rather than the messier ones we might
otherwise get.

This is something I tried before but wasn't useful then. It is now
- the interval replacement largely did the job that this is now doing,
but more expensively.