Numpy and Hypothesis idioms
np.prod is reduce-by-product, and avoids some imports.
Initialising an empty array is idiomatic as we're about to overwrite all
the zeros, and though irrelevant in this case can be faster for large
arrays.
Test with 'minimal' instead of 'find' - it's the internal equivalent,
with nice settings management and timeouts.