Rationalize which version of Python 3 we use
Depending on the phase of the task being run and the phase of the
moon, we were previously using one of Python 3.4.3, Python 3.5.1 or
Python 3.6.0 for our Python 3. This is silly and we should stop.

This updates everything to run on Python 3.6.1. It also uses tox's
default of the same interpeter as tox launched with, which means
we have to specify the Python version in fewer places.