Give tools virtualenv a unique name per reqs
Apparently git is very timestamp happy and keeps touching tools.txt
even if it hasn't changed (and doesn't set mtime based on commits
like I thought it did). This means that the makefile repeatedly
ends up blowing away our tools virtualenv, which takes a long time
to reinstall.

This changes that approach by giving the tools virtualenv a name
based on the hash of the current contents of requirements/tools.txt
and removing the time dependency. That means that every time the
file actually changes we'll get a new tools virtualenv, but as long
as its contents remain the same we'll keep the same stable name and
not have to rebuild.