Do not seed random.module from data by default
The motivation for seeding the random module in tests (which remains)
is that this is a great source of people writing tests with
nondeterminism that is not under Hypothesis's control and then
blaming Hypothesis for the inevitable confusing behaviour. We used
to have a health check that complained about that, but it basically
just annoyed people because most of the time this was happening
deep in the bowels of something they were depending on.

On the other hand, doing this for every test causes a lot of annoyances
too - it makes examples look distinct when they're not, slows things
down, and generally is a constant small annoyance that can trip us
up.

So instead we implement a convenient halfway house: If we're not asked
to seed the random module explicitly, we just seed it to zero. This is
fast and avoids all the above problems while getting most of the benefits.

If a user wants to seed the random module after all, the random_module
strategy remains available for their use.

Nominally this should require a documentation update, but we (I) never
actually updated the documtation to say that random\module was
no longer necessary anyway. Oops.