And so it has come to this
This adds a separate task for deployment and uses the Travis
API to determine when all the other jobs are complete, or fail
as soon as one of the others has failed.

This allows us to get the correct deployment semantics of only
deploying when all the other jobs have passed.

I'm not especially proud of this approach.