Make sure we pass through processed arguments
process_arguments_to_given mutates `arguments` and `kwargs`, but when we rearranged @given in a5c1200, the processed versions stayed within the
function, so the call to execute_explicit_examples would use the original
versions.  This changed the behaviour of @given and introduced problems
with pytest fixtures.

This patch returns the modified values from process_arguments_to_given.