Better lexicographic minimization
This takes a less scattershot approach to lexicographic
minimization and tries to do a reasonably small number of tests
instead of just brute forcing all the things. In many cases this
takes the number of test cases we run down from O(n) to O(log(n))
and in the remaining it takes it down to a better constant factor.

This introduces a new family of benchmarks, the lower bound benchmarks.
It does not compare them with the results of those benchmarks
before, for the very simple reason that those benchmarks did not
run in any sort of feasible time before.