Add test for missing line in extract_lambda_source
This runed out to be quite tricky, as the incremental parsing loop is
very particular for this edge case to be triggered.

First it needs to start with the lambda keyword to satisfy the if
statement, but then it has to continually fail for the for loops else
to be triggered. This is because the break statement will skip any for
loop else statement.

But it turns out that a comment in the middle of a lambda accomplish
this, which also proves how much of an edge case this truly is.

Relevant snippet from internal/reflection

```python
for i in hrange(len(source) - 1, len('lambda'), -1):
    prefix = source[:i]
    if 'lambda' not in prefix:
        return if_confused
    try:
        tree = ast.parse(prefix)
        source = prefix
        break
    except SyntaxError:
        continue
else:
    return if_confused
```