Refactor charmap caching [Fixes #767]
If the cached pickle of the charmap data was corrupted in any way
initialisation would fail.

Now if a cache file cannot be found or read, it will be
regenerated. The logic is also slightly refactored, so we do not
need to write a file to disk, only to immediately read it back.

This will also help #768 by not depending on a file being written
to disk.