Fix bug where mocks can be used as test runners
Fix bug where mocks can be used as test runners under certain
conditions. Specifically, if a mock is injected into a test via pytest
fixtures or patch decorators, and that mock is the first argument in the
list, hypothesis will think it represents  and turns the mock
into a test runner.  If this happens, the affected test always passes
because the mock is executed instead of the test body.

Specifics:
- Add a function to check for mocks by looking for random attributes
  that would be EXTREMELY unlikely to occur in an object that isn't a
  mock.
- Add a check in  to determine if the argument we think is
   is really a mock.  If that's the case, don't use it.