Avoid collisions in django models generator with field names
The first positional argument to the model generator (``model``) is
the reference to the model to work on. The remaining keyword arguments
passed refer to fields in the django model the caller intends to
override a specific strategy during sampling.

This commit renames the naming of this positional argument to allow
passing the field name 'model' as a keywork argument. Thi prevents the
TypeError raised as a result of multiple values passed with those
additional keywork arguments.