Improve behaviour of deadlines on the boundary
Deadlines currently interact poorly with shrinking in the
shrinker will try to find an example which is right on the
boundary of failing, which means that given the slightly
unreliable nature of timing, often when you replay the
example it doesn't fail! This results in a confusing user
experience.

The fix is twofold:

1. Give a better error message when this happens.
2. Try to avoid it happening in the first place by raising
   the deadline temporarily during the shrinking process, so
   that we stop at a boundary that is significantly above that
   of the deadline.

This fixes #892.