Dummy commit to create empty PR. WILL BE REMOVED
The purpose of this commit is to make it possible to create
a (n ultra-early WIP) PR on this branch.

No work has been done on this branch *yet*. The idea is to follow
DRMacIver's suggestion to kick off a discussion about the possible
approaches to the interface and implementation of the feature.